Move workflow into new workflows module.
